fix single parsnwir issue; move randomNumberGenerator to utils directory
