+++ src/services/helpers/hash.js
@@ -1,7 +1,7 @@
 const errors = require('@feathersjs/errors');
 const bcrypt = require('bcryptjs');
 const { userModel } = require('../user/model');
-const { getRandomInt } = require('../../helper/randomNumberGenerator');
+const { getRandomInt } = require('../../utils/randomNumberGenerator');
 
 const rndChar = () => {
 	const chars = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'R', 'S', 'T', 'U', 'V', 'W', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9'];
